requirements.txt
PyQt
cryptography